Ooof!  Replace the fasta sequence loading mechamism with one that is extensible to read from any source.  While doing that, greatly simplify the former fastastream (it took too many inputs) and generally clean up in that area.
Several bugs/features were fixed along the way, found mostly
by running the regression tests.

+++ kmer/libbio/mers.h
@@ -37,7 +37,7 @@ stringToMer(u32bit ms, char *str) {
 
   for (u32bit i=0; i<ms; i++) {
     mer <<= 2;
-    mer  |= compressSymbol[(int)str[i]];
+    mer  |= compressSymbol[str[i]];
   }
 
   return(mer);
