+++ kmer/snapper/encodedQuery.C
@@ -32,8 +32,8 @@ encodedQuery::encodedQuery(char const           *seq,
       substring <<= 2;
       substring  &= mermask;
 
-      if (validSymbol[(int)seq[seqLen - 1 - i]]) {
-        substring |= compressSymbol[ (int)complementSymbol[ (int)seq[seqLen - 1 - i] ]];
+      if (validSymbol[seq[seqLen - 1 - i]]) {
+        substring |= compressSymbol[ complementSymbol[ seq[seqLen - 1 - i] ]];
         timeUntilValid--;
       } else {
         timeUntilValid = k;
@@ -52,8 +52,8 @@ encodedQuery::encodedQuery(char const           *seq,
       substring <<= 2;
       substring  &= mermask;
 
-      if (validSymbol[(int)seq[i]]) {
-        substring |= compressSymbol[ (int)seq[i] ];
+      if (validSymbol[seq[i]]) {
+        substring |= compressSymbol[seq[i]];
         timeUntilValid--;
       } else {
         timeUntilValid = k;
